# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Friend Zone Timeline Feed API
  version: 0.1.0
  description: Timeline Feed first version with JWT auth, follow graph, post publishing, inbox fanout, and cursor pagination.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
  /auth/register:
    post:
      security: []
      summary: Register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Created
  /auth/login:
    post:
      security: []
      summary: Login and receive a JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
  /follows/{followee_id}:
    post:
      summary: Follow a user
      parameters:
        - $ref: '#/components/parameters/FolloweeID'
      responses:
        '200':
          description: OK
    delete:
      summary: Unfollow a user
      parameters:
        - $ref: '#/components/parameters/FolloweeID'
      responses:
        '200':
          description: OK
  /posts:
    post:
      summary: Publish a text post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePostRequest'
      responses:
        '201':
          description: Created
  /posts/{content_id}:
    get:
      summary: Get a post
      parameters:
        - $ref: '#/components/parameters/ContentID'
      responses:
        '200':
          description: OK
    delete:
      summary: Logically delete own post
      parameters:
        - $ref: '#/components/parameters/ContentID'
      responses:
        '200':
          description: OK
  /feed/timeline:
    get:
      summary: Read Timeline Feed
      description: >
        Use direction=latest for initial load, direction=newer with the top cursor for pull-to-refresh,
        and direction=older with the bottom cursor for infinite scroll.
      parameters:
        - name: direction
          in: query
          schema:
            type: string
            enum: [latest, newer, older]
            default: latest
        - name: cursor
          in: query
          schema:
            type: string
          description: Opaque cursor returned by a previous response.
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: OK
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    FolloweeID:
      name: followee_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ContentID:
      name: content_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  schemas:
    RegisterRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 64
        password:
          type: string
          minLength: 6
          maxLength: 128
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    CreatePostRequest:
      type: object
      required: [content_text]
      properties:
        content_text:
          type: string
          maxLength: 2000
    TimelineItem:
      type: object
      properties:
        content_id:
          type: string
          example: "330414614359576576"
        author_id:
          type: string
          example: "330414398441000960"
        content_text:
          type: string
        publish_time:
          type: string
          format: date-time